requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
Django==4.2.18
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
joblib==1.4.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.4.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
protobuf==3.19.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
statsmodels==0.14.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
